Add MakeProtective method for creating a Protective MBR.
In almost all cases, when creating new partition tables, a user should
create an GPT Table.  GPT requires a Protective MBR.
Add the MakeProtective method to the MBR interface so that someone
can easiliy create a valid Protective MBR.

A change is made to Check() to allow TYPE_GPT partitions to
be exactly 2TiB in length even though it is strictly outside the
MBR specification.  All tested gpt partitioning tools create a
partition of length 4294967295 (0xFFFFFFFF).